+++ client/modules/cm-tarnation/src/parser.ts
@@ -11,8 +11,8 @@ import {
 } from "@lezer/common"
 import { LanguageDescription, ParseContext } from "@wikijump/codemirror/cm"
 import { perfy } from "@wikijump/util"
-import { ChunkBuffer } from "./buffer"
-import { compileChunks } from "./chunk-parsing"
+import { ChunkBuffer } from "./chunk/buffer"
+import { compileChunks } from "./chunk/parsing"
 import { Nesting } from "./enums"
 import type { GrammarState } from "./grammar/state"
 import type { TarnationLanguage } from "./language"
